+++ R/utils.R
@@ -136,7 +136,7 @@ safeListAdd <- function(x, value) {
 }
 
 # clip of fixed length, zeropads if needed and deals with edge case
-clipAroundPeak <- function(wave, length) {
+clipAroundPeak <- function(wave, length, noise=FALSE) {
     if(length(wave) < length) {
         return(c(wave, rep(0, length - length(wave))))
     }
@@ -148,6 +148,13 @@ clipAroundPeak <- function(wave, length) {
     }
     if(high > length(wave)) {
         return(wave[(length(wave)-length+1):length(wave)])
+    }
+    if(noise) {
+        if(low - length > 1) {
+            return(wave[(low-length):(low-1)])
+        } else {
+            return(wave[1:length])
+        }
     }
     wave[low:high]
 }